api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pgvector
pydantic
python-multipart
Pillow
httpx
pytest
google-generativeai
